Revert "REFACTOR npm scripts & minor code corrections"
