+++ test/solidity/localDecoder.js
@@ -6,8 +6,8 @@ var miscLocal = require('./contracts/miscLocal')
 var structArrayLocal = require('./contracts/structArrayLocal')
 var VM = require('ethereumjs-vm')
 var utileth = require('ethereumjs-util')
-var Web3Providers = require('../../src/web3Provider/web3Providers')
-var util = require('../../src/helpers/global')
+var Web3Providers = require('../../babelify-src/web3Provider/web3Providers')
+var util = require('../../babelify-src/helpers/global')
 var intLocalTest = require('./localsTests/int')
 var miscLocalTest = require('./localsTests/misc')
 var misc2LocalTest = require('./localsTests/misc2')
